Extracts project stats to separate component
Moves project statistics calculations and display into a dedicated component.

This improves code organization and reusability by isolating the statistics logic. It also allows for conditional rendering of the stats based on screen size.
